gcd/python/requirements.txt
mysql-connector-python==8.0.33
matplotlib==3.7.1
pandas==2.0.1
dateparser==1.1.8
